// src/lib/server/archai/types.gen.ts
export type SingleGenerationResponse = {
    /**
     * URL of the generated image (or video, for /animate)
     */
    output: string;
    /**
     * Remaining account balance after this request
     */
    balance: number;
    /**
     * Cost charged for this request
     */
    cost: number;
};

export type ArrayGenerationResponse = {
    /**
     * List of generated image URLs
     */
    output: Array<string>;
    /**
     * Remaining account balance after this request
     */
    balance: number;
    /**
     * Cost charged for this request
     */
    cost: number;
};